Turned components into hooks and moved into their own files
